+++ packages/ipns/src/ipns/resolver.ts
@@ -1,4 +1,4 @@
-import { NotFoundError, isPeerId, isPublicKey } from '@libp2p/interface'
+import { isPeerId, isPublicKey } from '@libp2p/interface'
 import { multihashToIPNSRoutingKey, unmarshalIPNSRecord } from 'ipns'
 import { ipnsSelector } from 'ipns/selector'
 import { ipnsValidator } from 'ipns/validator'
@@ -7,7 +7,7 @@ import { base58btc } from 'multiformats/bases/base58'
 import { CID } from 'multiformats/cid'
 import * as Digest from 'multiformats/hashes/digest'
 import { DEFAULT_TTL_NS } from '../constants.ts'
-import { InvalidValueError, RecordsFailedValidationError, UnsupportedMultibasePrefixError, UnsupportedMultihashCodecError } from '../errors.js'
+import { InvalidValueError, RecordNotFoundError, RecordsFailedValidationError, UnsupportedMultibasePrefixError, UnsupportedMultihashCodecError } from '../errors.js'
 import { isCodec, IDENTITY_CODEC, SHA2_256_CODEC, isLibp2pCID } from '../utils.js'
 import type { IPNSResolveResult, ResolveOptions, ResolveResult } from '../index.js'
 import type { LocalStore } from '../local-store.js'
@@ -166,12 +166,13 @@ export class IPNSResolver {
     }
 
     if (options.offline === true) {
-      throw new NotFoundError('Record was not present in the cache or has expired')
+      throw new RecordNotFoundError('Record was not present in the cache or has expired')
     }
 
     this.log('did not have record locally')
 
     let foundInvalid = 0
+    const errors: Error[] = []
 
     await Promise.all(
       this.routers.map(async (router) => {
@@ -183,7 +184,8 @@ export class IPNSResolver {
             validate: false
           })
         } catch (err: any) {
-          this.log.error('error finding IPNS record - %e', err)
+          this.log.error('error finding IPNS record using router %s - %e', router.toString(), err)
+          errors.push(err)
 
           return
         }
@@ -195,7 +197,7 @@ export class IPNSResolver {
         } catch (err) {
           // we found a record, but the validator rejected it
           foundInvalid++
-          this.log.error('error finding IPNS record - %e', err)
+          this.log.error('error validating IPNS record from router %s - %e', router.toString(), err)
         }
       })
     )
@@ -205,7 +207,7 @@ export class IPNSResolver {
         throw new RecordsFailedValidationError(`${foundInvalid > 1 ? `${foundInvalid} records` : 'Record'} found for routing key ${foundInvalid > 1 ? 'were' : 'was'} invalid`)
       }
 
-      throw new NotFoundError('Could not find record for routing key')
+      throw new RecordNotFoundError('Could not find record for routing key')
     }
 
     const record = records[ipnsSelector(routingKey, records)]