+++ packages/bitswap/CHANGELOG.md
@@ -1,5 +1,41 @@
 # Changelog
 
+## [3.0.13](https://github.com/ipfs/helia/compare/bitswap-v3.0.12...bitswap-v3.0.13) (2025-12-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/utils bumped from ^2.3.2 to ^2.3.3
+
+## [3.0.12](https://github.com/ipfs/helia/compare/bitswap-v3.0.11...bitswap-v3.0.12) (2025-12-12)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/utils bumped from ^2.3.1 to ^2.3.2
+
+## [3.0.11](https://github.com/ipfs/helia/compare/bitswap-v3.0.10...bitswap-v3.0.11) (2025-12-12)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/utils bumped from ^2.3.0 to ^2.3.1
+
+## [3.0.10](https://github.com/ipfs/helia/compare/bitswap-v3.0.9...bitswap-v3.0.10) (2025-11-20)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/utils bumped from ^2.2.5 to ^2.3.0
+
 ## [3.0.9](https://github.com/ipfs/helia/compare/bitswap-v3.0.8...bitswap-v3.0.9) (2025-11-14)